# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api documentation for social media",
        "version": "1.0.0",
        "title": "Social media API",
        "contact": {
            "email": "sabin0kharel@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": ["http", "https"],
    "host": "localhost:5000",
    "basePath": "/",
    "securityDefinitions":{
        "APIKeyHeader":{
            "type": "apiKey",
            "in": "header",
            "name": "Token"
        }
    },
    "security": {
        "APIKeyHeader": []
    },
    "paths" : {
...
        "/post/": {
            "get": {
                "summary": "Get all Post",
                "description": "Get Post",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "tags": ["Post"],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Post"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/post/{id}":{
            "post": {
                "summary": "Create Post",
                "description": "Create Post",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "tags": ["Post"],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "Post object",
                    "required": true
                },{
                    "name": "body",
                    "in": "body",
                    "description": "Post object",
                    "required": true,
                    "schema": {
                        "type": "object",
                            "properties": {
                                "desc": {
                                    "type": "string"
                                },
                                "img": {
                                    "type": "string"
                                }
                            }
                        }
                    }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {"$ref": "#/definitions/Post"}
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        }
    }, 
    "definitions": {
...
        "Post": {
            "type": "object",
            "properties": {
                "_id": {
                     "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "img": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "likes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "InvalidResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}